requirements.txt
# Python dependencies
pandas>=2.0.0
pyarrow>=14.0.0
minio>=7.2.0
psycopg2-binary>=2.9.0
apache-airflow>=2.9.0
